+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Simplified `structuredClone` polyfill, avoided second tree pass in cases of transferring
 - Added support of `SuppressedError` to `structuredClone` polyfill
 - Removed unspecified unnecessary `ArrayBuffer` and `DataView` dependencies of `structuredClone` lack of which could cause errors in some entries in IE10-
-- Fixed handling of fractional number part and some special cases in [`Number.fromString`](https://github.com/tc39/proposal-number-fromstring)
+- Fixed handling of fractional number part in [`Number.fromString`](https://github.com/tc39/proposal-number-fromstring)
 - Compat data improvements:
   - Updated Opera Android 78 compat data mapping